// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-skill-details-model.ts
export class EmployeeSkillDetailsModel {
    employeeId: string;
    employeeSkillId:string;
    searchText:string;
    firstName: string;
    surName: string;
    userName: string;
    email: string;
    skillId: string;
    skillName: Date;
    monthsOfExperience: number;
    dateFrom: Date;
    dateTo: Date;
    comments: string;
    timeStamp: any;
    totalCount: number;
    isArchived: boolean;
    operationsPerformedBy:string;
    yearsOfExperience: any;
}